+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudFirmwareVersionObject.h
@@ -306,4 +306,10 @@
 /// 是否支持切换耳机仓鼠标起始点预设
 @property(nonatomic, readonly) BOOL allowSwitchMouseStartPointPresetForEarphoneCase;
 
+/// 是否支持AI问答模型选择
+@property(nonatomic, readonly) BOOL allowSelectAIModel;
+
+/// 穆斯林朝拜功能朝拜提醒是否支持开关控制
+@property(nonatomic, readonly) BOOL allowControlEnabledOfMuslimPrayerReminder;
+
 @end